server/requirements.txt
Flask==2.0.2
Flask_Cors==3.0.10
nltk==3.6.5
python_magic==0.4.24
Whoosh==2.7.4
gunicorn==20.1.0
